[CHORE] 불필요한 jpa 디렉토리 삭제
