--- src/main/kotlin/com/jaramgroupware/gateway/domain/jpa/serviceInfo/ServiceInfoRepository.kt
@@ -1,7 +1,0 @@
-package com.jaramgroupware.gateway.domain.jpa.serviceInfo
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ServiceInfoRepository: ReactiveCrudRepository<ServiceInfo, Int>,ServiceInfoCustomRepository
