+++ src/main/kotlin/com/jaramgroupware/gateway/domain/member/MemberCustomRepository.kt
@@ -1,4 +1,4 @@
-package com.jaramgroupware.gateway.domain.jpa.member
+package com.jaramgroupware.gateway.domain.member
 
 import reactor.core.publisher.Mono
