+++ src/main/kotlin/com/jaramgroupware/gateway/domain/member/MemberCustomRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.jaramgroupware.gateway.domain.jpa.member
+package com.jaramgroupware.gateway.domain.member
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.r2dbc.core.DatabaseClient
