+++ src/main/kotlin/com/jaramgroupware/gateway/domain/serviceInfo/ServiceInfo.kt
@@ -1,4 +1,4 @@
-package com.jaramgroupware.gateway.domain.jpa.serviceInfo
+package com.jaramgroupware.gateway.domain.serviceInfo
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
