+++ src/main/kotlin/com/jaramgroupware/gateway/domain/serviceInfo/ServiceInfoMapper.kt
@@ -1,4 +1,4 @@
-package com.jaramgroupware.gateway.domain.jpa.serviceInfo
+package com.jaramgroupware.gateway.domain.serviceInfo
 
 import io.r2dbc.spi.Row
 import io.r2dbc.spi.RowMetadata
