+++ src/main/kotlin/com/jaramgroupware/gateway/domain/serviceInfo/ServiceInfoRepository.kt
@@ -1,0 +1,7 @@
+package com.jaramgroupware.gateway.domain.serviceInfo
+
+import org.springframework.data.repository.reactive.ReactiveCrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ServiceInfoRepository: ReactiveCrudRepository<ServiceInfo, Int>, ServiceInfoCustomRepository
