+++ src/main/kotlin/com/jaramgroupware/gateway/service/MemberService.kt
@@ -1,7 +1,7 @@
 package com.jaramgroupware.gateway.service
 
-import com.jaramgroupware.gateway.domain.jpa.member.Member
-import com.jaramgroupware.gateway.domain.jpa.member.MemberRepository
+import com.jaramgroupware.gateway.domain.member.Member
+import com.jaramgroupware.gateway.domain.member.MemberRepository
 import com.jaramgroupware.gateway.dto.member.MemberResponseDto
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
